Move SaveLockFile to Context
