+++ Asterism/UpdateCommand.cs
@@ -51,8 +51,8 @@ internal class UpdateCommand {
 
         rootModule.CreatePropertySheet(true, librariesForConfigurations);
 
-        resolver.Dependencies = modules.ToList();
-        resolver.SaveLockFile();
+        context.Dependencies = modules.ToList();
+        context.SaveLockFile();
 
         return 0;
     }
